[issue-213] Removed silk and also the approximate commute handler works sooo much faster!
