+++ cocoon/survey/cocoon_algorithm/rent_algorithm.py
@@ -10,6 +10,7 @@ from cocoon.survey.cocoon_algorithm.weighted_scoring_algorithm import WeightScor
 
 # Import Commutes modules
 from cocoon.commutes.models import CommuteType
+from cocoon.commutes.models import ZipCodeBase, ZipCodeChild, CommuteType
 
 # Import DistanceWrapper
 from cocoon.commutes.distance_matrix.distance_wrapper import Distance_Matrix_Exception
@@ -18,6 +19,7 @@ from cocoon.commutes.distance_matrix.commute_retriever import retrieve_exact_com
 
 # Import Constants from commute module
 from cocoon.commutes.constants import CommuteAccuracy
+from cocoon.survey.constants import AVERAGE_BICYCLING_SPEED, AVERAGE_WALKING_SPEED, EXTRA_DISTANCE_LAT_LNG_APPROX
 
 # Import Geolocator
 import cocoon.houseDatabase.maps_requester as geolocator
@@ -104,10 +106,95 @@ class RentAlgorithm(SortingAlgorithms, WeightScoringAlgorithm, PriceAlgorithm, C
                 else:
                     lat_lng = (lat_lng_result[0], lat_lng_result[1])
 
-            for home in self.homes:
-                if not home.populate_approx_commutes(home.home, destination, lat_lng_dest=lat_lng):
-                    home.eliminate_home()
+            self.populate_approx_commutes(self.homes, destination, lat_lng_dest=lat_lng)
         print("STEP 2.2: time elapsed: {:.2f}s".format(time.time() - start_time))
+
+    def populate_approx_commutes(self, homes, destination, lat_lng_dest=""):
+        """
+        Based on the commute type of the destination, this function determines the algorithm method that will
+            be used to generate the approximation
+        :param home: (RentDatabaseModel) -> The home that the user is computing for
+        :param destination: (DestinationModel): The destination as a RentingDestinationsModel object
+        :param lat_lng_dest: ((decimal, decimal)): -> A Tuple of (latitude, longitude) for the destination
+        :return (Boolean): True if a valid pair match is found, False otherwise.
+        """
+        if destination.commute_type.commute_type == CommuteType.DRIVING:
+            return self.zip_code_approximation(homes, destination)
+        elif destination.commute_type.commute_type == CommuteType.TRANSIT:
+            return self.zip_code_approximation(homes, destination)
+        elif destination.commute_type.commute_type == CommuteType.BICYCLING:
+            return self.lat_lng_approximation(homes, destination, lat_lng_dest, AVERAGE_BICYCLING_SPEED)
+        elif destination.commute_type.commute_type == CommuteType.WALKING:
+            return self.lat_lng_approximation(homes, destination, lat_lng_dest, AVERAGE_WALKING_SPEED)
+        else:
+            return True
+
+    def lat_lng_approximation(self, home, destination, lat_lng_dest, average_speed):
+        """
+        This function given a home and a destination will determine the distance between the two homes based off of the
+            lat and lng points. Then once the distance is determined, then the commute time is determined based off of
+            the average speed.
+        :param home: (RentDatabaseModel) -> The home that the user is computing for
+        :param destination: (DestinationModel): The destination as a RentingDestinationsModel object
+        :param lat_lng_dest: ((decimal, decimal)): -> A Tuple of (latitude, longitude) for the destination
+        :param average_speed: (int) -> The average speed in mph that the person moves for the given mode of transport
+        :return: (Boolean): -> True: The home approximation was found and added
+                               False: The home was not able to have a approximation created
+        """
+        # Stores the lat and lng points for the home
+        lat_lng_home = (home.latitude, home.longitude)
+
+        # Returns the distance from the two lat lng points in miles
+        distance = geopy.distance.geodesic(lat_lng_home, lat_lng_dest).miles
+
+        # If the distance is less than a mile then don't add any distance since it is already so close
+        if distance > 1:
+            # Extra distance is determined by giving more distance to homes farther away
+            extra_distance = EXTRA_DISTANCE_LAT_LNG_APPROX * (1 - 1.0/distance)
+            # This normalizes the value since walking needs less of a weight than biking since homes
+            #   are more direct when walking.
+            distance += extra_distance * average_speed/AVERAGE_BICYCLING_SPEED
+        if average_speed is not 0:
+            commute_time_hours = distance / average_speed
+            commute_time = commute_time_hours * 60
+        else:
+            return False
+        self.approx_commute_times[destination] = commute_time
+        return True
+
+    def zip_code_approximation(self, homes, destination):
+        """
+        This is the zip_code_approximation algorithm. This assumes that the zip-code cache is already updated
+            and that all the valid pairs are already generated. This just goes through and finds the valid pairs
+            for the given zip_code and the destination
+        :param origin_zip: (string) -> The zip code of the origin, i.e home
+        :param destination: (DestinationModel): The destination as a RentingDestinationsModel object
+        :return (Boolean): True if a valid pair exists, False otherwise.
+        """
+        dest_zips = ZipCodeBase.objects.filter(zip_code__exact=destination.zip_code)
+        if dest_zips.exists():
+            for dest_zip in dest_zips:
+                child_zips = ZipCodeChild.objects.filter(base_zip_code=dest_zip). \
+                    filter(commute_type=destination.commute_type). \
+                    values_list('zip_code', 'commute_time_seconds')
+                for home in homes:
+                    home_exist = False
+                    home_commute_time = 0
+                    for zip_code, commute_time in child_zips:
+                        if home.home.zip_code in zip_code:
+                            home_exist = True
+                            home_commute_time = commute_time
+                            break
+                    if not home_exist:
+                        home.eliminate_home()
+                    else:
+                        home.approx_commute_times[destination] = home_commute_time/60
+
+
+
+
+        else:
+            return False
 
     def retrieve_exact_commutes(self):
         """
@@ -118,7 +205,7 @@ class RentAlgorithm(SortingAlgorithms, WeightScoringAlgorithm, PriceAlgorithm, C
             try:
                 # map list of HomeScore objects to full addresses
                 origin_addresses = list(map(lambda house:house.home.full_address,
-                                       self.homes[:number_of_exact_commutes_computed]))
+                                        self.homes[:number_of_exact_commutes_computed]))
 
                 destination_address = destination.full_address
